Update by commit

+++ cnn/predict.py
@@ -19,7 +19,7 @@ def run_model(model, itt, batch):
         batch.append([-1, "", [], [], ""])
     batch.sort(key = lambda x: -len(x[2]))
     xc, xw = batchify(*zip(*[(x[2], x[3]) for x in batch]), max(KERNEL_SIZES))
-    result = model(LongTensor(xc), LongTensor(xw))
+    result = model(xc, xw)
     for i in range(batch_size):
         y = itt[result[i].argmax()]
         p = result[i].max().exp().item()
